+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- [...]
+
+# v11.2.3 (16/09/2019)
+
 - **[UPDATE]** Add highlight animation on `Tabs`
 
 # v11.2.2 (05/09/2019)